+++ Pharmacy/Pharmacy.Application.Services.Data/Repositories/ProductTypeRepository.cs
@@ -66,7 +66,7 @@ namespace Pharmacy.Infrastructure.Data.Repositories
         public void Delete(int id)
         {
             var model = db.Find<ProductType>(id);
-            db.Remove(model);
+            if(model != null) db.Remove(model);
         }
     }
 }
