+++ Pharmacy/Pharmacy.Application.Services.Data/Repositories/SalesInfoRepository.cs
@@ -73,7 +73,8 @@ namespace Pharmacy.Infrastructure.Data.Repositories
         public void Delete(int id)
         {
             var model = db.Find<SalesInfo>(id);
-            db.Remove(model);
+            if (model != null)
+                db.Remove(model);
         }
     }
 }
